+++ api-contract/vue-ui/index.d.ts
@@ -1,11 +1,19 @@
 import * as vue from 'vue';
 import { DefineComponent, PropType, VNodeChild, VNode } from 'vue';
-import { GridOptions, Grid, Row, HeaderContext, CellContext, Column, CellFillOptions, ClipboardPasteResult, ClipboardPasteOptions, ExportFile } from '@open-grid/core';
+import { GridOptions, Grid, Row, Column, HeaderContext, CellContext, CellFillOptions, ClipboardPasteResult, ClipboardPasteOptions, ExportFile } from '@open-grid/core';
 export { AccessorColumnOptions, AccessorFnColumnDef, AccessorKey, AccessorKeyColumnDef, AnyColumnDef, CellContext, CellCoordinate, CellEditEvent, CellEditEventParams, CellEditHistoryAction, CellEditHistoryState, CellEditOption, CellEditParserContext, CellEditPhase, CellEditValidationContext, CellEditValidationResult, CellEditValidationState, CellEditingState, CellFillOptions, CellInteractionEvent, CellInteractionEventParams, CellRange, CellRangeSelectionState, ClipboardCellContext, ClipboardCopyOptions, ClipboardPasteCellContext, ClipboardPasteCommittedCell, ClipboardPasteOptions, ClipboardPasteResult, ClipboardPasteSkippedCell, ClipboardPasteSkippedReason, ClipboardPasteValidationError, Column, ColumnDef, ColumnFilter, ColumnFiltersState, ColumnHelper, ColumnId, ColumnMovePosition, ColumnOrderState, ColumnPinningPosition, ColumnPinningState, ColumnResizeEvent, ColumnResizeEventParams, ColumnResizePhase, ColumnSizingState, ColumnVisibilityState, DisplayColumnDef, ExpandedState, ExportFile, ExportFileOptions, FilterFn, FitColumnsToWidthOptions, Grid, GridCacheDiagnostics, GridCacheDiagnosticsEntry, GridCacheKey, GridOptions, GridState, GroupColumnDef, GroupingState, Header, HeaderContext, HeaderGroup, MoveFocusOptions, PaginationState, Row, RowId, RowInteractionEvent, RowInteractionEventParams, RowModel, RowSelectionCleanupScope, RowSelectionState, SortFn, SortingRule, SortingState, Updater, fitColumnsToWidth } from '@open-grid/core';
 import { ColumnVirtualizationPrimitiveOptions, GridLocalizationOverrides, GridDensity, RowVirtualizationPrimitiveOptions, GridLocalization } from '@open-grid/primitives';
 export { DEFAULT_GRID_LOCALIZATION, GRID_PREFERENCES_VERSION, GridDensity, GridLocalization, GridLocalizationOverrides, GridPreferences, GridPreferencesOptions, GridPreferencesState, GridPreferencesStorageEnvironmentLike, GridPreferencesStorageLike, createGridLocalization, createGridPreferences, getBrowserGridPreferencesStorage, parseGridPreferences, readGridPreferences, removeGridPreferences, serializeGridPreferences, writeGridPreferences } from '@open-grid/primitives';
 export { createColumnHelper, useGrid } from '@open-grid/vue';
 
+interface DataGridRenderContext<TData> {
+    grid: Grid<TData>;
+    rows: readonly Row<TData>[];
+    visibleColumns: readonly Column<TData, unknown>[];
+}
+interface DataGridErrorRenderContext<TData> extends DataGridRenderContext<TData> {
+    retry: (() => void) | undefined;
+}
 interface DataGridProps<TData> {
     ariaLabel?: string;
     localization?: GridLocalizationOverrides;
@@ -16,6 +24,12 @@ interface DataGridProps<TData> {
     onRetry?: () => void;
     loading?: boolean;
     loadingState?: VNodeChild;
+    renderToolbar?: (context: DataGridRenderContext<TData>) => VNodeChild;
+    renderEmptyState?: (context: DataGridRenderContext<TData>) => VNodeChild;
+    renderLoadingState?: (context: DataGridRenderContext<TData>) => VNodeChild;
+    renderErrorState?: (context: DataGridErrorRenderContext<TData>) => VNodeChild;
+    renderHeader?: (context: HeaderContext<TData, unknown>) => VNodeChild;
+    renderCell?: (context: CellContext<TData, unknown>) => VNodeChild;
     onGridReady?: GridReadyHandler<TData>;
     getRowClassName?: (row: Row<TData>) => string | undefined;
     getHeaderClassName?: (context: HeaderContext<TData, unknown>) => string | undefined;
@@ -116,6 +130,30 @@ declare const DataGrid: DefineComponent<vue.ExtractPropTypes<{
     };
     loadingState: {
         type: PropType<VNodeChild>;
+        default: undefined;
+    };
+    renderToolbar: {
+        type: PropType<(context: DataGridRenderContext<unknown>) => VNodeChild>;
+        default: undefined;
+    };
+    renderEmptyState: {
+        type: PropType<(context: DataGridRenderContext<unknown>) => VNodeChild>;
+        default: undefined;
+    };
+    renderLoadingState: {
+        type: PropType<(context: DataGridRenderContext<unknown>) => VNodeChild>;
+        default: undefined;
+    };
+    renderErrorState: {
+        type: PropType<(context: DataGridErrorRenderContext<unknown>) => VNodeChild>;
+        default: undefined;
+    };
+    renderHeader: {
+        type: PropType<(context: HeaderContext<unknown, unknown>) => VNodeChild>;
+        default: undefined;
+    };
+    renderCell: {
+        type: PropType<(context: CellContext<unknown, unknown>) => VNodeChild>;
         default: undefined;
     };
     onGridReady: {
@@ -249,6 +287,30 @@ declare const DataGrid: DefineComponent<vue.ExtractPropTypes<{
         type: PropType<VNodeChild>;
         default: undefined;
     };
+    renderToolbar: {
+        type: PropType<(context: DataGridRenderContext<unknown>) => VNodeChild>;
+        default: undefined;
+    };
+    renderEmptyState: {
+        type: PropType<(context: DataGridRenderContext<unknown>) => VNodeChild>;
+        default: undefined;
+    };
+    renderLoadingState: {
+        type: PropType<(context: DataGridRenderContext<unknown>) => VNodeChild>;
+        default: undefined;
+    };
+    renderErrorState: {
+        type: PropType<(context: DataGridErrorRenderContext<unknown>) => VNodeChild>;
+        default: undefined;
+    };
+    renderHeader: {
+        type: PropType<(context: HeaderContext<unknown, unknown>) => VNodeChild>;
+        default: undefined;
+    };
+    renderCell: {
+        type: PropType<(context: CellContext<unknown, unknown>) => VNodeChild>;
+        default: undefined;
+    };
     onGridReady: {
         type: PropType<GridReadyHandler<unknown>>;
         default: undefined;
@@ -350,6 +412,12 @@ declare const DataGrid: DefineComponent<vue.ExtractPropTypes<{
     onRetry: () => void;
     loading: boolean;
     loadingState: VNodeChild;
+    renderToolbar: (context: DataGridRenderContext<unknown>) => VNodeChild;
+    renderEmptyState: (context: DataGridRenderContext<unknown>) => VNodeChild;
+    renderLoadingState: (context: DataGridRenderContext<unknown>) => VNodeChild;
+    renderErrorState: (context: DataGridErrorRenderContext<unknown>) => VNodeChild;
+    renderHeader: (context: HeaderContext<unknown, unknown>) => VNodeChild;
+    renderCell: (context: CellContext<unknown, unknown>) => VNodeChild;
     onGridReady: GridReadyHandler<unknown>;
     getRowClassName: (row: Row<unknown>) => string | undefined;
     getHeaderClassName: (context: HeaderContext<unknown, unknown>) => string | undefined;
@@ -377,4 +445,4 @@ declare const DataGrid: DefineComponent<vue.ExtractPropTypes<{
 declare function createDataGrid<TData>(): DataGridComponent<TData>;
 declare function downloadExportFile(file: ExportFile): boolean;
 
-export { type ColumnVirtualizationOptions, DataGrid, type DataGridComponent, type DataGridProps, type GridReadyHandler, type HeaderActionMenuActionItem, type HeaderActionMenuContext, type HeaderActionMenuCustomItem, type HeaderActionMenuItem, type HeaderActionMenuItems, type HeaderActionMenuLabelItem, type HeaderActionMenuSeparatorItem, type RowVirtualizationOptions, createDataGrid, downloadExportFile };
+export { type ColumnVirtualizationOptions, DataGrid, type DataGridComponent, type DataGridErrorRenderContext, type DataGridProps, type DataGridRenderContext, type GridReadyHandler, type HeaderActionMenuActionItem, type HeaderActionMenuContext, type HeaderActionMenuCustomItem, type HeaderActionMenuItem, type HeaderActionMenuItems, type HeaderActionMenuLabelItem, type HeaderActionMenuSeparatorItem, type RowVirtualizationOptions, createDataGrid, downloadExportFile };